User Interface

This page contains all relevant docstrings for users to run a simulation.

LightMatter.run_simulation — Function

run_simulation(sys::Dict{String,Union{Expr,Vector{Expr}}}, initialtemps::Dict{String, Float64},
tspan::Tuple{Float64,Float64}, sim::Simulation; 
save, tolerance, max_step, min_step, callbacks)

Generates the problem the dynamics will solve and then solves the coupled system of ODE's.
Currently always uses Tsit5 for the integration routine but in the future that may be user-defined

# Arguments
- 'sys': Dictionary of ODE equations to be propagated
- 'initialtemps': Dictionary of initial temperatures of the bath
- 'tspan': Tuple of values for the dynamics to run between (the laser is centred on 0.0)
- 'sim': Simulation settings and parameters

# KWARGS
- Any key-word arguemnts from DiffEq that work with ODEProblems can be including in a namedtuple here

# Returns
- The solution of the dynamics calculation

LightMatter.post_production — Function

post_production(sol,file_name::String,initial_temps::Dict{String,Float64},output::Symbol,sim::Simulation)

Handles processing, saving the simulation after it has completed. Uses HDF5 file format
Currently the only output setting supported is :minimum which only saves the parameters, chemical potential and
progated systems

# Arguments
- 'sol': The solution of the simulation
- 'file_name': The name the user wishes the file to be saved as 
- 'initial_temps': Initial temperature the baths have been set to
- 'output': The output method required - may change in future
- 'sim': Simulation settings and parameters

# Returns
- Nothing is returned but a file is created